requirements.txt
pandas
kafka-python
streamlit
plotly
tqdm
